+++ extensions/elk-layout/main.js
@@ -1,7 +1,7 @@
 app.extension({
   id: 'examples.elk-layout',
   name: 'ELK Layout',
-  version: '1.3.1'
+  version: '1.3.2'
 });
 
 var config = app.assets.json('data/defaults.json');